Trait soco::algorithms::online::OnlineAlgorithm

pub trait OnlineAlgorithm<'a, T, P, M, O, C, D>: Fn(Online<P>, i32, &Schedule<T>, M, O) -> Result<Step<T, M>> + Send + Sync where
    T: Value<'a>,
    P: Problem<T, C, D> + 'a,
    M: Memory<'a, T, P, C, D>,
    O: Options<T, P, C, D>,
    C: ModelOutputSuccess,
    D: ModelOutputFailure, {
    fn next(
        &self, 
        o: Online<P>, 
        xs: &Schedule<T>, 
        prev_m_: Option<M>, 
        options: O
    ) -> Result<Step<T, M>> { ... }
    fn next_with_default_options(
        &self, 
        o: Online<P>, 
        xs: &Schedule<T>, 
        prev_m_: Option<M>
    ) -> Result<Step<T, M>> { ... }
}

Implementation of an online algorithm.

P - problem

M - memory

O - options

Receives the arguments:

o - online problem instance

t - current time slot

xs - schedule up to the previous time slot

prev_m - latest memory, is the default if nothing was memorized

options - algorithm options

Provided Methods

fn next(
&self,
o: Online<P>,
xs: &Schedule<T>,
prev_m_: Option<M>,
options: O
) -> Result<Step<T, M>>

Executes the next iteration of an online algorithm.

fn next_with_default_options(
&self,
o: Online<P>,
xs: &Schedule<T>,
prev_m_: Option<M>
) -> Result<Step<T, M>>

Executes the next iteration of an online algorithm with default options.