Module soco::cost

Definition of cost functions.

Structs

Cost

Result of cost computation.

CostFn

Cost functions that arrived over time. Individual cost functions may have different domains. For example, in a predictive online setting, a cost function arriving at time $t$ generally has the domain $[t, t + w]$.

SingleCostFn

Cost function (from time $t_start$).

Traits

DecisionSpace

Abstract definition of a decision space which is defined by which points lie within and which points lie outside it.

Type Definitions

FailableCost

A cost with associated information in case of a failure.

FailableCostFn

A cost function returning costs with associated information in case of a failure.

RawCost

A cost without associated information.

RawCostFn

A cost function returning costs without any associated information.

RawSingleCostFn

A single cost function returning costs without any associated information.