Module soco::model

Abstract and concrete definitions of models that are used to generate problem instances.

Modules

data_center

Utilities to model the cost of data centers.

Enums

ModelOutput

Output of a model.

Traits

Model

Model which is used to generate problem instances and update them online.

ModelOutputFailure

Failures of a model.

ModelOutputSuccess

Results of a model.

OfflineInput

Model inputs to generate offline problem.

OnlineInput

Model inputs to update a problem instance (online) to the next time slot. Encapsulates information for the current time slot as well as all time slots in the prediction window.

Functions

verify_update

Utility to verify that the update of an online instance is valid.