Function soco::model::data_center::loads::apply_predicted_loads

pub fn apply_predicted_loads<'a, 'b, T>(
    d: i32, 
    e: i32, 
    objective: impl Fn(i32, &Config<T>, &LoadProfile, &LoadFractions) -> IntermediateObjective + Send + Sync + 'b, 
    predicted_loads: Vec<PredictedLoadProfile>, 
    t_start: i32
) -> SingleCostFn<'b, Config<T>, DataCenterModelOutputSuccess, DataCenterModelOutputFailure> where
    T: Value<'a>, 

Optimally applies loads to a model to obtain a cost function.

$d$ - number of dimensions

$e$ - number of job types

$objective$ - cost function to minimize w.r.t. load assignments

$predicted_loads$ - vector of predicted loads for all time slots that should be supported by the returned cost function

$t_start$ - time offset, i.e. time of first load samples