Trait soco::model::Model

pub trait Model<T, P, A, B, C, D>: Clone + Send + Sync where
    P: Problem<T, C, D>,
    A: OfflineInput,
    B: OnlineInput,
    C: ModelOutputSuccess,
    D: ModelOutputFailure, {
    fn to(&self, input: A) -> P;
    fn update(&self, o: &mut Online<P>, input: B);
}

Model which is used to generate problem instances and update them online.

Required Methods

fn to(&self, input: A) -> P

Generates an offline problem instance given some $input$ (with certainty).

fn update(&self, o: &mut Online<P>, input: B)

Performs an online update of the given problem instance $o$ with some $input$ (which may be uncertain).