Function soco::streaming::offline::solve

pub fn solve<'a, T, R, P, O, A, B, C, D>(
    model: &'a impl Model<T, P, A, B, C, D>, 
    alg: &impl OfflineAlgorithm<T, R, P, O, C, D>, 
    options: O, 
    offline_options: OfflineOptions, 
    input: A
) -> Result<(Schedule<T>, Cost<C, D>, u128)> where
    T: Value<'a>,
    R: OfflineResult<T>,
    P: Problem<T, C, D> + 'a,
    O: Options<T, P, C, D> + 'a,
    A: OfflineInput,
    B: OnlineInput,
    C: ModelOutputSuccess,
    D: ModelOutputFailure, 

Generates problem instance from model and solves it using an offline algorithm.