Function soco::streaming::online::start

pub fn start<T, P, M, O, A, B, C, D>(
    addr: SocketAddr, 
    model: impl Model<T, P, A, B, C, D> + 'static, 
    alg: &'static impl OnlineAlgorithm<'static, T, P, M, O, C, D>, 
    options: O, 
    w: i32, 
    input: A, 
    sender: Option<Sender<&'static str>>
) -> Result<OfflineResponse<T, C, D, M>> where
    T: Value<'static>,
    P: Problem<T, C, D> + 'static,
    M: Memory<'static, T, P, C, D>,
    O: Options<T, P, C, D> + 'static,
    A: OfflineInput,
    B: OnlineInput,
    C: ModelOutputSuccess,
    D: ModelOutputFailure, 

Generates problem instance from model and streams online algorithm using the provided input. Then, starts backend. Returns initial schedule, initial integral schedule, and latest memory of the algorithm.